+++ R/PKNCA.options.R
@@ -49,6 +49,18 @@
       return("lin up/log down")
     match.arg(x, c("lin up/log down", "linear"))
   },
+  conc_above=function(x, default=FALSE, description=FALSE) {
+    if (description) {
+      return(
+        "For the time_above parameter, what concentration should the value be above?"
+      )
+    }
+    if (default)
+      return(NA_real_)
+    stopifnot("conc_above must be a scalar"=length(x) == 1)
+    stopifnot("conc_above must be numeric"=is.numeric(x))
+    x
+  },
   conc.na=function(x, default=FALSE, description=FALSE) {
     if (description)
       return(paste(
@@ -295,7 +307,7 @@
 #' included in the summary.
 #' 
 #' See the vignette 'Options for Controlling PKNCA' for a current list
-#' of options.
+#' of options (\code{vignette("Options-for-Controlling-PKNCA", package="PKNCA")}).
 #' 
 #' @param \dots options to set or get the value for
 #' @param default (re)sets all default options