+++ R/class-PKNCAdose.R
@@ -136,8 +136,9 @@ PKNCAdose.data.frame <- function(data, formula, route, rate, duration,
 #' @param ... Arguments passed to another setRoute function
 #' @return The object with an updated route
 #' @export
-setRoute <- function(object, ...)
+setRoute <- function(object, ...) {
   UseMethod("setRoute")
+}
 #' @rdname setRoute
 #' @export
 setRoute.PKNCAdose <- function(object, route, ...) {
@@ -247,6 +248,12 @@ getGroups.PKNCAdose <- function(...) {
   getGroups.PKNCAconc(...)
 }
 
+#' @describeIn group_vars.PKNCAconc Get group_vars for a PKNCAdose object
+#' @exportS3Method dplyr::group_vars
+group_vars.PKNCAdose <- function(x) {
+  group_vars.PKNCAconc(x)
+}
+
 #' @rdname getData.PKNCAconc
 #' @export
 #' @importFrom nlme getData
@@ -298,7 +305,3 @@ print.PKNCAdose <- function(x, n=6, summarize=FALSE, ...) {
 #' @rdname print.PKNCAconc
 #' @export
 summary.PKNCAdose <- summary.PKNCAconc
-
-#' @rdname split.PKNCAconc
-#' @export
-split.PKNCAdose <- split.PKNCAconc